Move authz validation to grants and use traits
