+++ lib/Server/Grants/Interfaces/OidcCapableGrantTypeInterface.php
@@ -4,5 +4,4 @@ namespace SimpleSAML\Modules\OpenIDConnect\Server\Grants\Interfaces;
 
 interface OidcCapableGrantTypeInterface
 {
-
 }
